main-app/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
psycopg2-binary==2.9.3
redis==4.5.1
grpcio==1.50.0
grpcio-tools==1.50.0
boto3==1.26.0
